Fix for #7286
git-svn-id: https://svn.participatoryculture.org/svn/dtv/trunk@4661 bce20531-a64c-4166-8316-56baf4a01729

Original sha1: 054b0bb82da79c05ed0122639a07cfa4476678d6

+++ tv/portable/dl_daemon/bittorrentdtv.py
@@ -381,7 +381,7 @@ class TorrentDownload:
             try:
                 fast_resume_data = {
                     'already_got': storagewrapper.get_have_list(),
-                    'mtimes': dict([(f, path.getmtime(f)) for \
+                    'mtimes': dict([(f, long(path.getmtime(f))) for \
                             f, size in files]),
                 }
                 self.fast_resume_queue.put(bencode(fast_resume_data))
